python/rag-indexer/requirements.txt
openai>=1.0.0
azure-search-documents>=11.4.0
requests>=2.31.0
